+++ src/BusinessLogic/Security/IPasswordHasher.cs
@@ -1,0 +1,7 @@
+namespace BusinessLogic.Security
+{
+    public interface IPasswordHasher
+    {
+        byte[] Hash(string username, string password);
+    }
+}
