+++ src/BusinessLogic/Security/IPasswordPolicyValidator.cs
@@ -1,0 +1,9 @@
+using DataAccess.Entities;
+
+namespace BusinessLogic.Security
+{
+    public interface IPasswordPolicyValidator
+    {
+        void Validate(string password, string username, Persona persona, PoliticaSeguridad politica);
+    }
+}
