+++ src/Presentation/UserForm.cs
@@ -6,36 +6,38 @@ namespace Presentation
 {
     public partial class UserForm : Form
     {
-        private readonly IUserService _userService;
+        private readonly IUserAuthenticationService _authService;
+        private readonly IUserManagementService _managementService;
         private readonly string _username;
 
-        public UserForm(IUserService userService, string username)
+        public UserForm(IUserAuthenticationService authService, IUserManagementService managementService, string username)
         {
             InitializeComponent();
-            _userService = userService;
+            _authService = authService;
+            _managementService = managementService;
             _username = username;
 
             btnCambiarContrasena.Click += (s, e) =>
             {
-                var form = new CambioContrasenaForm(_userService, _username);
+                var form = new CambioContrasenaForm(_authService, _username);
                 form.ShowDialog();
             };
 
             btnCambiarPreguntas.Click += (s, e) =>
             {
-                var form = new PreguntasSeguridadForm(_userService, _username);
+                var form = new PreguntasSeguridadForm(_authService, _username);
                 form.ShowDialog();
             };
 
             btnMiPerfil.Click += (s, e) =>
             {
-                var user = _userService.GetUserByUsername(_username);
+                var user = _managementService.GetUserByUsername(_username);
                 if (user != null)
                 {
-                    var persona = _userService.GetPersonaById(user.IdPersona);
+                    var persona = _managementService.GetPersonaById(user.IdPersona);
                     if (persona != null)
                     {
-                        var form = new ProfileForm(user, persona);
+                        var form = new ProfileForm(user, persona); // This form also needs refactoring
                         form.ShowDialog();
                     }
                     else
